wip(pulumi-aws): split lambda edge definition
